+++ Entregable 2/Src/LibraryModelado/Item/ButtonGoToPage.cs
@@ -31,7 +31,7 @@ namespace Proyecto.Item
         /// <param name="image">Imagen del boton.</param>
         /// <param name="color">Color del boton en Hexadecimal.</param>
         /// <param name="pageName">Pagina para mostrar.</param>
-        public ButtonGoToPage(string name, Space level, int positionX, int positionY, int width, int height, string image, string color, string pageName)
+        public ButtonGoToPage(string name, Space level, float positionX, float positionY, float width, float height, string image, string color, string pageName)
         : base(name, level, positionX, positionY, width, height, image)
         {
             this.Color = color;
